refactor(core, web): 패키지 경로 수정 및 사용하지 않는 클래스 제거
